fixed rep

+++ BlueJFILES/Main.java
@@ -573,7 +573,7 @@ public class Main {
     public static void updateBoard(Human human, String[] textArea, int currentPlayer)
     { 
        board.setTurnCount("Turn: " + (economy.getTurns() ));
-       board.setRepCount("Rep: " + (human.getReputation()) + "%" );
+       board.setRepCount("Rep: " + (int) (human.getReputation()) + "%" );
        board.setMoneyCount("Money: " + human.getMoney() );
        board.setTextArea("-"+ textArea[0] + "\n" + "-" + textArea[1] + "\n" + "-" + textArea[2] + "\n"+ "-" + textArea[3] + "\n" + "-" + textArea[4] );
        board.updateTurnBorder(currentPlayer);
